完成我的设置模块的个人信息相关的绑定(手机，邮箱，第三方等)，以及整合短信、邮件、站内通知。基本完成商城推广位的申请

+++ src/main/java/com/cn/leedane/mybatis/SqlProvider.java
@@ -374,7 +374,7 @@ public class SqlProvider {
 		return executeSQL(sql, params);
 	}
 	
-	public String getObjectCreateUserId(String tableName, int tableId){
+	public String getObjectCreateUserId(String tableName, long tableId){
 		return "select create_user_id from "+tableName +" where id="+tableId;
 	}
